pyproject.toml
[build-system]
requires = ["setuptools>=60.5", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "colint"
version = "0.2.1"
description = ""
authors = [
    { name = "Davide Colì", email = "davide.coli@secomind.com" },
    { name = "Michele Puppin", email = "michele.puppin@secomind.com" }
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "black>=25.1.0",
    "exceptiongroup>=1.2.2",
    "flake8-docstrings>=1.7.0",
    "GitPython>=3.1.43",
    "ipykernel>=6.29.5",
    "isort>=6.0.1",
    "tokenize-rt>=6.0.0",
    "toml>=0.10.2",
    "tomli>=2.0.1",
    "typing-extensions>=4.12.2",
]

[project.scripts]
colint = "colint.colinter:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
"colint" = ["*.toml"]
